+++ nativelink-util/BUILD.bazel
@@ -21,12 +21,12 @@ rust_library(
         "src/lib.rs",
         "src/metrics_utils.rs",
         "src/platform_properties.rs",
+        "src/proto_stream_utils.rs",
         "src/resource_info.rs",
         "src/retry.rs",
         "src/store_trait.rs",
         "src/tls_utils.rs",
         "src/write_counter.rs",
-        "src/write_request_stream_wrapper.rs",
     ],
     proc_macro_deps = [
         "@crates//:async-trait",
@@ -67,6 +67,7 @@ rust_test_suite(
         "tests/fastcdc_test.rs",
         "tests/fs_test.rs",
         "tests/health_utils_test.rs",
+        "tests/proto_stream_utils_test.rs",
         "tests/resource_info_test.rs",
         "tests/retry_test.rs",
     ],
@@ -80,14 +81,17 @@ rust_test_suite(
         ":nativelink-util",
         "//nativelink-config",
         "//nativelink-error",
+        "//nativelink-proto",
         "@crates//:bytes",
         "@crates//:futures",
         "@crates//:hex",
         "@crates//:mock_instant",
+        "@crates//:parking_lot",
         "@crates//:pretty_assertions",
         "@crates//:rand",
         "@crates//:sha2",
         "@crates//:tokio",
+        "@crates//:tokio-stream",
         "@crates//:tokio-util",
     ],
 )
