+++ nativelink-util/src/lib.rs
@@ -23,9 +23,9 @@ pub mod grpc_utils;
 pub mod health_utils;
 pub mod metrics_utils;
 pub mod platform_properties;
+pub mod proto_stream_utils;
 pub mod resource_info;
 pub mod retry;
 pub mod store_trait;
 pub mod tls_utils;
 pub mod write_counter;
-pub mod write_request_stream_wrapper;
